requirements.txt
# Core
pandas>=2.0
numpy>=1.24
pydantic>=2.0
pyyaml>=6.0
python-dotenv>=1.0

# Machine Learning
scikit-learn>=1.3
xgboost>=2.0
lightgbm>=4.0
optuna>=3.5

# Explainability
shap>=0.44

# Visualization
plotly>=5.18
seaborn>=0.13
matplotlib>=3.7

# Web Dashboard
flask>=3.0

# LLM Providers (install as needed)
# openai>=1.0
# anthropic>=0.25
# google-generativeai>=0.5
